push

+++ RestoranTablesService/Classes/Drink.cs
@@ -10,10 +10,10 @@ namespace RestaurantTablesService.Classes
     {
         public int DrinkID { get; set; }
         public string DrinkName { get; set; }
-        public double DrinkPrice { get; set; }
-        public double DrinkPrimePrice { get; set; }
+        public decimal DrinkPrice { get; set; }
+        public decimal DrinkPrimePrice { get; set; }
 
-        public Drink(int drinkID, string drinkName, double drinkPrice, double drinkPrimePrice)
+        public Drink(int drinkID, string drinkName, decimal drinkPrice, decimal drinkPrimePrice)
         {
             DrinkID = drinkID;
             DrinkName = drinkName;
